requirements.txt
requests
websockets
python-dotenv
urllib3<2
pytest
pytest-asyncio
flake8
black
isort
autoflake
